// src/app/shared/constants/enums/controls/form/form-component-type.enum.ts
/**
 * Form component type enum for form components
 */
export enum ComponentType {
  INPUT = 'app-input',
  TEXTAREA = 'app-textarea',
  TEXT_EDITOR = 'app-text-editor',
  CHECKBOX = 'app-checkbox',
  CHECKBOX_LIST = 'app-checkbox-list',
  RADIO = 'app-radio',
  SHAPE = 'app-light-shape',
  BORDER = 'app-light-border',
  ICON = 'app-light-icon',
  SWITCH = 'app-switch',
  SELECT = 'app-select',
  COLOR_INPUT = 'app-color-input',
  IMAGE = 'app-imag',
  IMAGE_UPLOAD = 'app-image-upload',
  MULTI_IMAGE_UPLOAD = 'app-multi-image-upload',
  LOGO = 'app-logo',
  FILE_ATTACHMENT = 'app-file-attachment',
  DATE_PICKER = 'app-date-picker',
  DATE_TIME_PICKER = 'app-date-time-picker',
  TIME_PICKER = 'app-time-picker',
  RANGE = 'app-range',
  WEEK_SELECTOR = 'kt-week-selector',
  MENTION_EDITOR = 'kt-mention-editor',
  DATE_RANGE_PICKER = 'app-date-range-picker',
  LAYOUT = 'app-light-layout',
  AUTOCOMPLETE_ADDRESS = 'kt-autocomplete-address',
  TEMPLATE = 'template'
}